FE-Refactor: list item prop 수정 및 폴더 경로 변경
